@@ -1,15 +1,29 @@
 {
   "name": "bob-core",
-  "version": "1.0.1",
+  "version": "1.2.0",
   "description": "BOB Core",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "type": "module",
   "files": [
     "/dist"
   ],
+  "exports": {
+    ".": {
+      "import": "./dist/esm/index.js",
+      "require": "./dist/cjs/index.js"
+    }
+  },
+  "typesVersions": {
+    "*": {
+      "*": [
+        "./dist/cjs/index.d.ts"
+      ]
+    }
+  },
   "scripts": {
     "start": "node -r @swc-node/register debug/main.ts",
-    "build": "tsc",
+    "build": "rimraf ./dist && npm run build:cjs && npm run build:esm",
+    "build:cjs": "tsc --project tsconfig.json --module commonjs --verbatimModuleSyntax false --moduleResolution node10 --outDir ./dist/cjs && echo >./dist/cjs/package.json '{\"type\":\"commonjs\"}'",
+    "build:esm": "tsc --project tsconfig.json --outDir ./dist/esm",
     "prepare": "npm run build",
     "test": "jest"
   },
@@ -24,7 +38,8 @@
     "@types/string-similarity": "^4.0.2",
     "jest": "^29.7.0",
     "ts-jest": "^29.1.5",
-    "typescript": "^5.4.5"
+    "typescript": "^5.7.3",
+    "rimraf": "^6.0.1"
   },
   "dependencies": {
     "chalk": "^4.1.2",
