@@ -1,8 +1,9 @@
-module.exports = {
+export default {
     preset: 'ts-jest',
     testEnvironment: 'node',
     testPathIgnorePatterns: [
         "/node_modules/",
         "/dist/"
     ],
+    transform: {}
 };
