@@ -1,4 +1,4 @@
-import {BobError} from "./errors/BobError";
+import {BobError} from "@/src/errors/index.js";
 
 export class ExceptionHandler {
     handle(err: Error | BobError) {
