@@ -1,5 +1,5 @@
-export * from './Command';
-export * from './Cli';
-export * from './errors';
-export * from './contracts';
-export * from './options';
+export * from './Command.js';
+export * from './Cli.js';
+export * from './errors/index.js';
+export * from './contracts/index.js';
+export * from './options/index.js';
